quant-sweep: five blockers from review — PackedCandidate execution, no unsafe cast, real payload bytes, family_id thresholds, per-tensor limits
1. Replace FamilyCandidate tuple execution with PackedCandidate-ready types: packer
   closure now returns extra: Vec<u8> instead of Vec<f32> in all 5 families (NF4,
   SymInt4, INT8, Ternary, MixedTile). No more unsafe from_raw_parts byte cast
   in the runner.

2. Delete the unsafe _extra byte cast: runner now receives extra_bytes: Vec<u8>
   directly from the packer and passes &extra_bytes to the unpacker. MixedTile
   removed its f32 reinterpretation layer — the packer builds routing_table +
   f32 payload as Vec<u8> directly.

3. ByteAccounting from actual payload lengths: runner now computes ByteAccounting
   by measuring actual codes.len(), converting scales+biases to LE bytes, and
   passing extra_bytes.len(). No more estimator functions with zero-filled vectors.

4. Key thresholds by QuantFamilyId: max_weight_nrmse_by_family lookup now uses
   format!("{:?}", family_id) instead of fc.label, fixing the NRMSE gate for
   long-format family labels.

5. Per-tensor candidate counter: tensor_candidate_count is tracked per tensor
   loop with a break after max_candidates. The modulo trick is gone.

+++ compute-core/src/quantization/sweep/families/mod.rs
@@ -91,7 +91,8 @@ pub struct FamilyCandidate {
     /// Resolved parameter set for this candidate, as JSON Value.
     pub parameters: Value,
     /// Pack weights into (codes, scales, biases, extra).
-    pub packer: Box<dyn Fn(&[f32], usize, usize) -> (Vec<u8>, Vec<f32>, Vec<f32>, Vec<f32>) + Send + Sync>,
+    /// extra is already LE bytes (Vec<u8>), not Vec<f32>.
+    pub packer: Box<dyn Fn(&[f32], usize, usize) -> (Vec<u8>, Vec<f32>, Vec<f32>, Vec<u8>) + Send + Sync>,
     /// Reconstruct weights from packed representation.
     pub unpacker:
         Box<dyn Fn(&[u8], &[f32], &[f32], &[u8], usize, usize) -> Vec<f32> + Send + Sync>,
